Move @verbose annotation to base package
