+++ src/base/command.ts
@@ -7,6 +7,8 @@ import { Platform } from '../util/platform';
 
 const QUOTE = Platform.OS === 'win32' ? '"' : '\'';
 
+import { workspace } from 'vscode';
+
 export class CommandOption {
     protected privacy = false;
     constructor(public readonly name: string, public readonly value?: string, public readonly redacted = true, public readonly quoted = false) {
@@ -51,4 +53,24 @@ export class CommandText {
        this.options.push(option);
        return this;
     }
+}
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export function verbose(_: unknown, key: string, descriptor: TypedPropertyDescriptor<Function>): void {
+    let fnKey: string | undefined;
+    // eslint-disable-next-line @typescript-eslint/ban-types
+    let fn: Function | undefined;
+
+    if (typeof descriptor.value === 'function') {
+        fnKey = 'value';
+        fn = descriptor.value;
+    } else {
+        throw new Error('not supported');
+    }
+
+    descriptor[fnKey] = function(...args: unknown[]): unknown {
+        const v = workspace.getConfiguration('openshiftToolkit').get<number>('outputVerbosityLevel');
+        const command = fn.apply(this, args) as CommandText;
+        return v > 0 ? command.addOption(new CommandOption('-v', `${v}`)) : command;
+    };
 }
